requirements.txt
Flask==3.1.1
openai==1.99.5
python-dotenv==1.1.1
gunicorn==23.0.0